Sample of a py-to-exe converter inclusion
Folder 'Executables' is too big to be used as compacted file here due to the subfolders with the scripts used in the transactions of SEXTOW project. To adding some future features in format of scripts to the application Basic script, this file will be an auto convert the files to be used if correctly user. Please, open the file before executing it in CMD, because some changes must be done before running the converter, with information of current py file to be converted to exe.