Folder to put Intermediario Project inside
This folder has the auxiliar folders to script put temporary files, auxiliar versions of the project, versions to approve and fixed info tables to be useful through the VBA Script and possibly some future exe python file script. To make it work, just uncompact the folder, install the most recent version of the project and substitute the version used as sample (with the same name of the original project).